Numeric.Rounded.Hardware

Documentation

newtype Rounded (r :: RoundingMode) a

A type tagged with a rounding direction.

The rounding direction is effective for a single operation. You won't get the correctly-rounded result for a compound expression like (a - b * c) :: Rounded 'TowardInf Double.

In particular, a negative literal like -0.1 :: Rounded r Double doesn't yield the correctly-rounded value for -0.1. To get the correct value, call fromRational explicitly (i.e. fromRational (-0.1) :: Rounded r Double) or use NegativeLiterals extension.

Constructors

Rounded

Fields

getRounded :: a

data RoundingMode

The type for IEEE754 rounding-direction attributes.

Constructors

ToNearest
Round to the nearest value (IEEE754 roundTiesToEven)

TowardNegInf
Round downward (IEEE754 roundTowardNegative)

TowardInf
Round upward (IEEE754 roundTowardPositive)

TowardZero
Round toward zero (IEEE754 roundTowardZero)

class Rounding (r :: RoundingMode)

This class allows you to recover the runtime value from a type-level rounding mode.

See rounding.

Minimal complete definition

roundingT

class Ord a => RoundedRing a

Rounding-controlled version of Num.

Minimal complete definition

roundedAdd, roundedSub, roundedMul, roundedFusedMultiplyAdd, roundedFromInteger, backendNameT

class RoundedRing a => RoundedFractional a

Rounding-controlled version of Fractional.

Minimal complete definition

roundedDiv, roundedFromRational

class RoundedRing a => RoundedSqrt a

Rounding-controlled version of sqrt.

Minimal complete definition

roundedSqrt