Numeric.Rounded.Hardware.Backend.C

Description

The types in this module implements rounding-mode-controlled operations in C.

There are several ways to control rounding mode in C, and an appropriate technology will be selected at compile time. This library implements the following options:

C99 fesetround

On x86 systems,

SSE2 MXCSR (for Float and Double)

AVX512 EVEX encoding (for Float and Double)

x87 Control Word (for LongDouble)

On AArch64, FPCR

You should not need to import this module directly.

This module is not available if the package flag pure-hs is set.

Documentation

newtype CFloat

A wrapper providing particular instances for RoundedRing, RoundedFractional and RoundedSqrt.

This type is different from CFloat from Foreign.C.Types.

Constructors

CFloat Float

newtype CDouble

A wrapper providing particular instances for RoundedRing, RoundedFractional and RoundedSqrt.

This type is different from CDouble from Foreign.C.Types.

Constructors

CDouble Double

data family MVector s a

data family Vector a

roundedFloatFromInt64 :: RoundingMode -> Int64 -> Float

roundedFloatFromWord64 :: RoundingMode -> Word64 -> Float

roundedDoubleFromInt64 :: RoundingMode -> Int64 -> Double

roundedDoubleFromWord64 :: RoundingMode -> Word64 -> Double