Numeric.Rounded.Hardware.Internal

Documentation

>>> import Data.Int

binaryFloatToDecimalDigitsRn

Arguments

:: forall a. RealFloat a

=> RoundingMode
rounding mode

-> Int
prec

-> a
a non-negative number (zero, normal or subnormal)

-> ([Int], Int)

>>> binaryFloatToDecimalDigitsRn ToNearest 3 (0.125 :: Double)
([1,2,5],0)
>>> binaryFloatToDecimalDigitsRn ToNearest 3 (12.5 :: Double)
([1,2,5],2)

binaryFloatToFixedDecimalDigitsRn

Arguments

:: forall a. RealFloat a

=> RoundingMode
rounding mode

-> Int
prec

-> a
a non-negative number (zero, normal or subnormal)

-> [Int]

>>> binaryFloatToFixedDecimalDigitsRn ToNearest 3 (0.125 :: Double)
[1,2,5]
>>> binaryFloatToFixedDecimalDigitsRn ToNearest 3 (12.5 :: Double)
[1,2,5,0,0]

binaryFloatToDecimalDigits

Arguments

:: RealFloat a

=> a
a non-negative number (zero, normal or subnormal)

-> ([Int], Int)

>>> binaryFloatToDecimalDigits (0.125 :: Double)
([1,2,5],0)
>>> binaryFloatToDecimalDigits (12.5 :: Double)
([1,2,5],2)

showEFloatRn :: RealFloat a => RoundingMode -> Maybe Int -> a -> ShowS

>>> showEFloatRn ToNearest (Just 0) (0 :: Double) ""
"0e0"
>>> showEFloatRn ToNearest Nothing (0 :: Double) ""
"0.0e0"
>>> showEFloatRn ToNearest Nothing (0.5 :: Double) ""
"5.0e-1"

showFFloatRn :: RealFloat a => RoundingMode -> Maybe Int -> a -> ShowS

>>> showFFloatRn ToNearest (Just 0) (0 :: Double) ""
"0"
>>> showFFloatRn ToNearest Nothing (0 :: Double) ""
"0.0"
>>> showFFloatRn ToNearest Nothing (-0 :: Double) ""
"-0.0"
>>> showFFloatRn ToNearest Nothing (-0.5 :: Double) ""
"-0.5"

showGFloatRn :: RealFloat a => RoundingMode -> Maybe Int -> a -> ShowS

data RoundingMode

The type for IEEE754 rounding-direction attributes.

Constructors

ToNearest
Round to the nearest value (IEEE754 roundTiesToEven)

TowardNegInf
Round downward (IEEE754 roundTowardNegative)

TowardInf
Round upward (IEEE754 roundTowardPositive)

TowardZero
Round toward zero (IEEE754 roundTowardZero)

oppositeRoundingMode :: RoundingMode -> RoundingMode

Returns the opposite rounding direction.

TowardNegInf and TowardInf are swapped.

class Rounding (r :: RoundingMode)

This class allows you to recover the runtime value from a type-level rounding mode.

See rounding.

Minimal complete definition

roundingT

rounding :: Rounding r => proxy r -> RoundingMode

Recovers the value from type-level rounding mode.

reifyRounding :: RoundingMode -> (forall s. Rounding s => Proxy s -> a) -> a

Lifts a rounding mode to type-level.

newtype Rounded (r :: RoundingMode) a

A type tagged with a rounding direction.

The rounding direction is effective for a single operation. You won't get the correctly-rounded result for a compound expression like (a - b * c) :: Rounded 'TowardInf Double.

In particular, a negative literal like -0.1 :: Rounded r Double doesn't yield the correctly-rounded value for -0.1. To get the correct value, call fromRational explicitly (i.e. fromRational (-0.1) :: Rounded r Double) or use NegativeLiterals extension.

Constructors

Rounded

Fields

getRounded :: a

data family MVector s a

data family Vector a

nextUp :: RealFloat a => a -> a

IEEE 754 nextUp operation.

nextUp 1 == (0x1.000002p0 :: Float)

nextUp 1 == (0x1.0000_0000_0000_1p0 :: Double)

nextUp (1/0) == (1/0 :: Double)

nextUp (-1/0) == (- maxFinite :: Double)

nextUp 0 == (0x1p-1074 :: Double)

nextUp (-0) == (0x1p-1074 :: Double)

nextUp (-0x1p-1074) == (-0 :: Double)

isNegativeZero (nextUp (-0x1p-1074) :: Double)

nextDown :: RealFloat a => a -> a

IEEE 754 nextDown operation.

nextDown 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)

nextDown 1 == (0x1.fffffep-1 :: Float)

nextDown (1/0) == (maxFinite :: Double)

nextDown (-1/0) == (-1/0 :: Double)

nextDown 0 == (-0x1p-1074 :: Double)

nextDown (-0) == (-0x1p-1074 :: Double)

nextDown 0x1p-1074 == (0 :: Double)

nextDown 0x1p-1022 == (0x0.ffff_ffff_ffff_fp-1022 ::Double)

nextTowardZero :: RealFloat a => a -> a

nextTowardZero 1 == (0x1.ffff_ffff_ffff_fp-1 :: Double)

nextTowardZero 1 == (0x1.fffffep-1 :: Float)

nextTowardZero (1/0) == (maxFinite :: Double)

nextTowardZero (-1/0) == (-maxFinite :: Double)

nextTowardZero 0 == (0 :: Double)

isNegativeZero (nextTowardZero (-0 :: Double))

nextTowardZero 0x1p-1074 == (0 :: Double)

distanceUlp :: RealFloat a => a -> a -> Maybe Integer

fusedMultiplyAdd :: RealFloat a => a -> a -> a -> a

IEEE 754 fusedMultiplyAdd operation.

May make use of hardware FMA instructions if the target architecture has it; set fma3 package flag on x86 systems.

\(a :: Double) (b :: Double) (c :: Double) -> fusedMultiplyAdd a b c == fromRational (toRational a * toRational b + toRational c)

roundedFromInteger_default :: RealFloat a => RoundingMode -> Integer -> a

roundedFromRational_default :: RealFloat a => RoundingMode -> Rational -> a

intervalFromInteger_default :: RealFloat a => Integer -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalFromRational_default :: RealFloat a => Rational -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

class RealFloatConstants a where

Methods

positiveInfinity :: a

\(+\infty\)

negativeInfinity :: a

\(-\infty\)

maxFinite :: a

minPositive :: a

pi_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\pi\)

pi_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\pi\)

three_pi_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(3\pi\)

three_pi_up :: Rounded 'TowardInf a

The correctly-rounded value of \(3\pi\)

five_pi_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(5\pi\)

five_pi_up :: Rounded 'TowardInf a

The correctly-rounded value of \(5\pi\)

log2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\log_e 2\)

log2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\log_e 2\)

exp1_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\exp(1)\)

exp1_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\exp(1)\)

exp1_2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\exp(1/2)\)

exp1_2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\exp(1/2)\)

expm1_2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\exp(-1/2)\)

expm1_2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\exp(-1/2)\)

sqrt2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\sqrt{2}\)

sqrt2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\sqrt{2}\)

sqrt2m1_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(\sqrt{2}-1\)

sqrt2m1_up :: Rounded 'TowardInf a

The correctly-rounded value of \(\sqrt{2}-1\)

sqrt1_2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(1/\sqrt{2}\)

sqrt1_2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(1/\sqrt{2}\)

three_minus_2sqrt2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(3-2\sqrt{2}\)

three_minus_2sqrt2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(3-2\sqrt{2}\)

two_minus_sqrt2_down :: Rounded 'TowardNegInf a

The correctly-rounded value of \(2-\sqrt{2}\)

two_minus_sqrt2_up :: Rounded 'TowardInf a

The correctly-rounded value of \(2-\sqrt{2}\)

class (RoundedSqrt a, RoundedRing_Vector vector a) => RoundedSqrt_Vector vector a where

Lifted version of RoundedSqrt

Minimal complete definition

Nothing

Methods

map_roundedSqrt :: RoundingMode -> vector a -> vector a

Equivalent to map . roundedSqrt

default map_roundedSqrt :: Vector vector a => RoundingMode -> vector a -> vector a

class (RoundedFractional a, RoundedRing_Vector vector a) => RoundedFractional_Vector vector a where

Lifted version of RoundedFractional

Minimal complete definition

Nothing

Methods

zipWith_roundedDiv :: RoundingMode -> vector a -> vector a -> vector a

Equivalent to zipWith . roundedDiv

default zipWith_roundedDiv :: Vector vector a => RoundingMode -> vector a -> vector a -> vector a

class RoundedRing a => RoundedRing_Vector vector a where

Lifted version of RoundedRing

Minimal complete definition

Nothing

Methods

roundedSum :: RoundingMode -> vector a -> a

Equivalent to \r -> foldl (roundedAdd r) 0

default roundedSum :: (Vector vector a, Num a) => RoundingMode -> vector a -> a

zipWith_roundedAdd :: RoundingMode -> vector a -> vector a -> vector a

Equivalent to zipWith . roundedAdd

default zipWith_roundedAdd :: Vector vector a => RoundingMode -> vector a -> vector a -> vector a

zipWith_roundedSub :: RoundingMode -> vector a -> vector a -> vector a

Equivalent to zipWith . roundedSub

default zipWith_roundedSub :: Vector vector a => RoundingMode -> vector a -> vector a -> vector a

zipWith_roundedMul :: RoundingMode -> vector a -> vector a -> vector a

Equivalent to zipWith . roundedMul

default zipWith_roundedMul :: Vector vector a => RoundingMode -> vector a -> vector a -> vector a

zipWith3_roundedFusedMultiplyAdd :: RoundingMode -> vector a -> vector a -> vector a -> vector a

Equivalent to zipWith3 . roundedFusedMultiplyAdd

default zipWith3_roundedFusedMultiplyAdd :: Vector vector a => RoundingMode -> vector a -> vector a -> vector a -> vector a

class RoundedRing a => RoundedSqrt a where

Rounding-controlled version of sqrt.

Minimal complete definition

roundedSqrt

Methods

roundedSqrt :: RoundingMode -> a -> a

intervalSqrt :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

class RoundedRing a => RoundedFractional a where

Rounding-controlled version of Fractional.

Minimal complete definition

roundedDiv, roundedFromRational

Methods

roundedDiv :: RoundingMode -> a -> a -> a

roundedRecip :: RoundingMode -> a -> a

default roundedRecip :: Num a => RoundingMode -> a -> a

roundedFromRational :: RoundingMode -> Rational -> a

roundedFromRealFloat :: RealFloat b => RoundingMode -> b -> a

default roundedFromRealFloat :: (Fractional a, RealFloat b) => RoundingMode -> b -> a

intervalDiv :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalDivAdd :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalRecip :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalFromRational :: Rational -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

class Ord a => RoundedRing a where

Rounding-controlled version of Num.

Minimal complete definition

roundedAdd, roundedSub, roundedMul, roundedFusedMultiplyAdd, roundedFromInteger, backendNameT

Methods

roundedAdd :: RoundingMode -> a -> a -> a

roundedSub :: RoundingMode -> a -> a -> a

roundedMul :: RoundingMode -> a -> a -> a

roundedFusedMultiplyAdd :: RoundingMode -> a -> a -> a -> a

roundedFromInteger :: RoundingMode -> Integer -> a

intervalAdd :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

\x_lo x_hi y_lo y_hi -> intervalAdd (Rounded x_lo) (Rounded x_hi) (Rounded y_lo) (Rounded y_hi) == (Rounded (roundedAdd TowardNegInf x_lo y_lo), Rounded (roundedAdd TowardInf x_hi y_hi))

intervalSub :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

\x_lo x_hi y_lo y_hi -> intervalSub (Rounded x_lo) (Rounded x_hi) (Rounded y_lo) (Rounded y_hi) == (Rounded (roundedSub TowardNegInf x_lo y_hi), Rounded (roundedSub TowardInf x_hi y_lo))

intervalMul :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalMulAdd :: Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> Rounded 'TowardNegInf a -> Rounded 'TowardInf a -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

intervalFromInteger :: Integer -> (Rounded 'TowardNegInf a, Rounded 'TowardInf a)

backendNameT :: Tagged a String

backendName :: RoundedRing a => proxy a -> String

Returns the name of backend as a string.

Example:

>>> :m + Data.Proxy
>>> backendName (Proxy :: Proxy Double)
"FastFFI+SSE2"

data RoundingMode

The type for IEEE754 rounding-direction attributes.

Constructors

ToNearest
Round to the nearest value (IEEE754 roundTiesToEven)

TowardNegInf
Round downward (IEEE754 roundTowardNegative)

TowardInf
Round upward (IEEE754 roundTowardPositive)

TowardZero
Round toward zero (IEEE754 roundTowardZero)

oppositeRoundingMode :: RoundingMode -> RoundingMode

Returns the opposite rounding direction.

TowardNegInf and TowardInf are swapped.

class Rounding (r :: RoundingMode)

This class allows you to recover the runtime value from a type-level rounding mode.

See rounding.

Minimal complete definition

roundingT

rounding :: Rounding r => proxy r -> RoundingMode

Recovers the value from type-level rounding mode.

reifyRounding :: RoundingMode -> (forall s. Rounding s => Proxy s -> a) -> a

Lifts a rounding mode to type-level.

newtype Rounded (r :: RoundingMode) a

A type tagged with a rounding direction.

The rounding direction is effective for a single operation. You won't get the correctly-rounded result for a compound expression like (a - b * c) :: Rounded 'TowardInf Double.

In particular, a negative literal like -0.1 :: Rounded r Double doesn't yield the correctly-rounded value for -0.1. To get the correct value, call fromRational explicitly (i.e. fromRational (-0.1) :: Rounded r Double) or use NegativeLiterals extension.

Constructors

Rounded

Fields

getRounded :: a

data family MVector s a

data family Vector a